refactor: moved middleware
