+++ src/app.ts
@@ -1,35 +1,18 @@
 // import createError from 'http-errors';
-import express, { NextFunction } from 'express';
+import express from 'express';
 import swaggerUI from 'swagger-ui-express';
 import path from 'path';
-import fs from 'fs';
 import YAML from 'yamljs';
 // import HttpError from './utils/error/httpError';
 import userRouter from './resources/users/user.router';
 import boardsRouter from './resources/board/board.router';
-import { finished } from 'stream';
+import loggerMiddleware from './middleware/loggerMiddleware';
 // import sendHttpError from './middleware/sendHttpError'
 
 const app: express.Application = express();
 const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
 app.use(express.json());
 // app.use(sendHttpError());
-
-function loggerMiddleware(request: express.Request, response: express.Response, next: NextFunction) {
-  const { method, url } = request;
-  const start = Date.now();
-
-  next();
-
-  finished(response, () => {
-    const data = `${method} ${url} [${ Date.now() - start }ms] \n`;
-    fs.appendFile('log.txt', data, (e) => {
-      if (e) {
-        console.error(e);
-      }
-    })
-  });
-}
 
 
 app.use(loggerMiddleware);
